// src/app/models/Note.ts
export class Note {
    id_note: number | undefined;
    id_usuario: number | undefined;
    title_note: string | undefined;
    description_note: string | undefined;
    id_prioridade: number | undefined;
    nm_prioridade: string | undefined;
    create_at: Date | undefined;
    update_at: Date | undefined;
    sector: string | undefined;
    noteColor: string | undefined;
}